# internal/generator/apigen/testdata/golden/fixture-api-session/openapi.json
{
  "components": {
    "schemas": {
      "CreateTenantInput": {
        "properties": {
          "name": {
            "description": "An objects name",
            "type": "string"
          },
          "slug": {
            "description": "A URL friendly version of a string",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "type": "object"
      },
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResponseLinks": {
        "properties": {
          "first": {
            "nullable": true,
            "type": "string"
          },
          "last": {
            "nullable": true,
            "type": "string"
          },
          "next": {
            "nullable": true,
            "type": "string"
          },
          "prev": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResponseMeta": {
        "properties": {
          "requestId": {
            "description": "Unique identifier for the request",
            "type": "string"
          },
          "totalCount": {
            "description": "Total count of items for paginated responses",
            "nullable": true,
            "type": "integer"
          }
        },
        "required": [
          "requestId"
        ],
        "type": "object"
      },
      "TenantView": {
        "description": "Customer-facing projection of the Tenant table.",
        "properties": {
          "id": {
            "description": "UUID v4 with automatic base62 encoding for client-facing APIs",
            "type": "string"
          },
          "internalDebugLabel": {
            "type": "string"
          },
          "name": {
            "description": "An objects name",
            "type": "string"
          },
          "userCount": {
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "userCount",
          "internalDebugLabel"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API generated from the fixture-api schema",
    "title": "Fixture-Api API",
    "version": "__PSGEN_OPENAPI_VERSION__"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/auth/me": {
      "get": {
        "operationId": "SessionCurrentTenantHandler",
        "parameters": [
          {
            "description": "Optional request correlation ID echoed back in the response.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TenantView"
                    },
                    "links": {
                      "$ref": "#/components/schemas/ResponseLinks"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "currentTenant",
        "tags": [
          "session"
        ]
      }
    },
...
    "/api/tenants": {
      "get": {
        "operationId": "TenantListTenantsHandler",
        "parameters": [
          {
            "description": "Optional request correlation ID echoed back in the response.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/TenantView"
                      },
                      "type": "array"
                    },
                    "links": {
                      "$ref": "#/components/schemas/ResponseLinks"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "listTenants",
        "tags": [
          "tenant"
        ]
      },
      "post": {
        "operationId": "TenantCreateTenantHandler",
        "parameters": [
          {
            "description": "Optional request correlation ID echoed back in the response.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TenantView"
                    },
                    "links": {
                      "$ref": "#/components/schemas/ResponseLinks"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "createTenant",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/tenants/{id}": {
      "get": {
        "description": "Fetch one tenant by id.",
        "operationId": "TenantGetTenantHandler",
        "parameters": [
          {
            "description": "Optional request correlation ID echoed back in the response.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "id parameter",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "UUID v4 with automatic base62 encoding for client-facing APIs",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeArchived",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TenantView"
                    },
                    "links": {
                      "$ref": "#/components/schemas/ResponseLinks"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "getTenant",
        "tags": [
          "tenant"
        ]
      },
      "patch": {
        "operationId": "TenantUpdateSecretHandler",
        "parameters": [
          {
            "description": "Optional request correlation ID echoed back in the response.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "id parameter",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "UUID v4 with automatic base62 encoding for client-facing APIs",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "secret": {
                    "type": "string"
                  }
                },
                "required": [
                  "secret"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TenantView"
                    },
                    "links": {
                      "$ref": "#/components/schemas/ResponseLinks"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "updateSecret",
        "tags": [
          "tenant"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Runtime server",
      "url": "__PSGEN_OPENAPI_BASE_URL__"
    }
  ]
}