# internal/generator/apigen/testdata/golden/fixture-api-session/routes.go
// Code generated by psgen apigen. DO NOT EDIT.
//
// This file provides RegisterRoutes() to wire up API routes with your implementations.
// Import this module and call RegisterRoutes() with a chi router and Config struct.

package fixtureapi

import (
	"encoding/json"
	"fmt"
	gohttp "net/http"
	"strconv"
	"time"

	"github.com/go-chi/chi/v5"
	orm "github.com/parable-platform/platform-schemas/orm/fixture-db"
	types "github.com/parable-platform/platform-schemas/types/go/fixture-api"
	runtimemiddleware "github.com/parable-work/superschematic/runtime/http/go/middleware"
	runtimerouting "github.com/parable-work/superschematic/runtime/http/go/routing"
	runtimesession "github.com/parable-work/superschematic/runtime/http/go/session"
	"go.uber.org/zap"
)

const (
	defaultOpenAPIVersion = "1.0.0"
	defaultOpenAPIBaseURL = "http://localhost:8080"
)

// Config holds all dependencies needed to register API routes.
// Pass this to RegisterRoutes() to wire up your business logic.
type Config struct {
	// DB is the database connection for middleware operations (auth, tenant context, etc.)
	DB orm.DatabaseInterface

	// Logger is the base logger for request-scoped logging
	Logger *zap.Logger

	// OpenAPIVersion is injected into the served OpenAPI document.
	// Optional: defaults to "1.0.0" if empty.
	OpenAPIVersion string

	// OpenAPIBaseURL is injected into the served OpenAPI document servers list.
	// Optional: defaults to "http://localhost:8080" if empty.
	OpenAPIBaseURL string
	// AuthMiddleware is the authentication middleware to apply to protected routes.
	AuthMiddleware func(gohttp.Handler) gohttp.Handler

	// PayloadDecryptor decrypts encrypted endpoint request bodies.
	PayloadDecryptor runtimemiddleware.PayloadDecryptor

	// Implementations contains all namespace implementation instances
	Implementations Implementations
}

// Validate checks that all required dependencies are provided.
// Returns an error if any required field is nil.
func (c *Config) Validate() error {
	if c.DB == nil {
		return fmt.Errorf("Config.DB is required")
	}
	if c.Logger == nil {
		return fmt.Errorf("Config.Logger is required")
	}
	if c.AuthMiddleware == nil {
		return fmt.Errorf("Config.AuthMiddleware is required")
	}
	if c.PayloadDecryptor == nil {
		return fmt.Errorf("Config.PayloadDecryptor is required for encrypted endpoints")
	}
	if err := c.Implementations.ValidateImplementations(); err != nil {
		return err
	}
	return nil
}

// RegisterRoutes registers all API routes on the provided chi router.
// The router should already have base middleware applied (RequestID, RealIP, etc.).
//
// Example usage:
//
//	r := chi.NewRouter()
//	r.Use(middleware.RequestID)
//	r.Use(middleware.RealIP)
//	r.Use(middleware.Recoverer)
//
//	fixtureapi.RegisterRoutes(r, fixtureapi.Config{
//		DB:                          db,
//		Logger:                      logger,
//		AuthMiddleware:              authMiddleware,
//		PayloadDecryptor:            payloadDecryptor,
//		Implementations: fixtureapi.Implementations{
//			Session: &mySessionImpl{},
//			Tenant: &myTenantImpl{},
//		},
//	})
func RegisterRoutes(r chi.Router, cfg Config) error {
	// Validate configuration
	if err := cfg.Validate(); err != nil {
		return fmt.Errorf("invalid config: %w", err)
	}

	openAPIVersion := cfg.OpenAPIVersion
	if openAPIVersion == "" {
		openAPIVersion = defaultOpenAPIVersion
	}

	openAPIBaseURL := cfg.OpenAPIBaseURL
	if openAPIBaseURL == "" {
		openAPIBaseURL = defaultOpenAPIBaseURL
	}

	// Add logger middleware for request-scoped logging
	r.Use(LoggerMiddleware(cfg.Logger))

	// Add database to context for all requests
	r.Use(DatabaseMiddleware(cfg.DB))

	// Request completion log: one authoritative line per request with identity,
	// outcome, and performance fields. Sets http.route span attribute and records
	// OTel span error events for 5xx responses.
	r.Use(runtimemiddleware.RequestLogMiddleware(LoggerFromContext))

	runtimerouting.Register(r, utilityRoutes(openAPIVersion, openAPIBaseURL))

	// API routes
	r.Route("/api", func(r chi.Router) {
		runtimerouting.Register(r, publicAPIRoutes(cfg))
		runtimerouting.RegisterGroup(
			r,
			protectedAPIRoutes(cfg),
			cfg.AuthMiddleware,
		)
	})

	return nil
}

func utilityRoutes(apiVersion, baseURL string) []runtimerouting.Route {
	routes := []runtimerouting.Route{
		{Method: gohttp.MethodGet, Path: "/", Handler: createIndexHandler()},
		{Method: gohttp.MethodGet, Path: "/api/openapi.json", Handler: createOpenAPIHandler(apiVersion, baseURL)},
		{Method: gohttp.MethodGet, Path: "/api/docs", Handler: createRapiDocHandler()},
	}
	return routes
}

func publicAPIRoutes(cfg Config) []runtimerouting.Route {
	return []runtimerouting.Route{}
}

func protectedAPIRoutes(cfg Config) []runtimerouting.Route {
	return []runtimerouting.Route{
		{
			Method:  "GET",
			Path:    "/auth/me",
			Handler: createSessionCurrentTenantHandler(cfg.Implementations.Session),
		},
		{
			Method:  "GET",
			Path:    "/tenants",
			Handler: createTenantListTenantsHandler(cfg.Implementations.Tenant),
			Middlewares: []runtimerouting.Middleware{
				runtimemiddleware.RateLimit(60, time.Minute, LoggerFromContext),
				runtimemiddleware.BodyLimit(1, RespondError),
				runtimesession.RequirePermissions("tenants.read"),
			},
		},
		{
			Method:  "POST",
			Path:    "/tenants",
			Handler: createTenantCreateTenantHandler(cfg.Implementations.Tenant),
			Middlewares: []runtimerouting.Middleware{
				runtimemiddleware.DecryptPayloadMiddleware(cfg.PayloadDecryptor, LoggerFromContext),
				runtimesession.RequirePermissions("tenants.write"),
			},
		},
		// Fetch one tenant by id.
		{
			Method:  "GET",
			Path:    "/tenants/{id}",
			Handler: createTenantGetTenantHandler(cfg.Implementations.Tenant),
			Middlewares: []runtimerouting.Middleware{
				runtimemiddleware.RateLimit(60, time.Minute, LoggerFromContext),
				runtimemiddleware.BodyLimit(1, RespondError),
				runtimemiddleware.Timeout(5*time.Second, LoggerFromContext),
				runtimesession.RequirePermissions("tenants.read"),
			},
		},
		{
			Method:  "PATCH",
			Path:    "/tenants/{id}",
			Handler: createTenantUpdateSecretHandler(cfg.Implementations.Tenant),
			Middlewares: []runtimerouting.Middleware{
				runtimemiddleware.DecryptPayloadMiddleware(cfg.PayloadDecryptor, LoggerFromContext),
				runtimesession.RequirePermissions("tenants.write"),
			},
		},
	}
}

// =============================================================================
// Handler factory functions
// =============================================================================
// Each handler factory creates a handler that calls the corresponding
// implementation method. This allows for dependency injection and testing.

// createSessionCurrentTenantHandler creates a handler for GET /api/auth/me
func createSessionCurrentTenantHandler(impl SessionImplementation) gohttp.HandlerFunc {
	return func(w gohttp.ResponseWriter, r *gohttp.Request) {

		// Ensure request context is still valid before entering implementation logic.
		if err := CheckContext(r.Context()); err != nil {
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}

		// Call implementation
		result, err := impl.CurrentTenant(r.Context())
		if err != nil {
			// Get logger from context and use proper error handling
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}
		// Validate output (if validation method exists)
		if validator, ok := interface{}(result).(interface {
			Validate() interface{ HasErrors() bool }
		}); ok {
			if validationErrors := validator.Validate(); validationErrors.HasErrors() {
				RespondError(w, r, gohttp.StatusInternalServerError, "Invalid response from implementation")
				return
			}
		}

		// Respond with result
		RespondJSONEnvelope(w, gohttp.StatusOK, result, r)
	}
}

// createTenantCustomHandlerHandler creates a handler for POST /api/tenant/custom-handler
//
// This endpoint uses @manualRouteRegistration: it is NOT registered by
// RegisterRoutes. Register it yourself on the router.
func createTenantCustomHandlerHandler(impl TenantImplementation) gohttp.HandlerFunc {
	return func(w gohttp.ResponseWriter, r *gohttp.Request) {

		// Ensure request context is still valid before entering implementation logic.
		if err := CheckContext(r.Context()); err != nil {
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}

		// Call implementation
		result, err := impl.CustomHandler(r.Context())
		if err != nil {
			// Get logger from context and use proper error handling
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}
		// Validate output (if validation method exists)
		if validator, ok := interface{}(result).(interface {
			Validate() interface{ HasErrors() bool }
		}); ok {
			if validationErrors := validator.Validate(); validationErrors.HasErrors() {
				RespondError(w, r, gohttp.StatusInternalServerError, "Invalid response from implementation")
				return
			}
		}

		// Respond with result
		RespondJSONEnvelope(w, gohttp.StatusOK, result, r)
	}
}

// createTenantListTenantsHandler creates a handler for GET /api/tenants
func createTenantListTenantsHandler(impl TenantImplementation) gohttp.HandlerFunc {
	return func(w gohttp.ResponseWriter, r *gohttp.Request) {

		// Ensure request context is still valid before entering implementation logic.
		if err := CheckContext(r.Context()); err != nil {
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}

		// Call implementation
		result, err := impl.ListTenants(r.Context())
		if err != nil {
			// Get logger from context and use proper error handling
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}
		// Validate output array (if validation method exists on elements)
		for i, item := range result {
			if validator, ok := interface{}(item).(interface {
				Validate() interface{ HasErrors() bool }
			}); ok {
				if validationErrors := validator.Validate(); validationErrors.HasErrors() {
					RespondError(w, r, gohttp.StatusInternalServerError, "Invalid response from implementation at index "+strconv.Itoa(i))
					return
				}
			}
		}

		// Respond with result
		RespondCollectionEnvelope(w, gohttp.StatusOK, result, r)
	}
}

// createTenantCreateTenantHandler creates a handler for POST /api/tenants
func createTenantCreateTenantHandler(impl TenantImplementation) gohttp.HandlerFunc {
	return func(w gohttp.ResponseWriter, r *gohttp.Request) {
		// Parse and validate input
		var input types.CreateTenantInput
		var rawInput json.RawMessage
		if err := json.NewDecoder(r.Body).Decode(&rawInput); err != nil {
			RespondError(w, r, gohttp.StatusBadRequest, "Invalid request body")
			return
		}
		if string(rawInput) == "null" {
			RespondError(w, r, gohttp.StatusBadRequest, "input is required")
			return
		}
		if err := json.Unmarshal(rawInput, &input); err != nil {
			RespondError(w, r, gohttp.StatusBadRequest, "Invalid request body")
			return
		}

		// Validate input
		if validationErrors := input.Validate(); validationErrors.HasErrors() {
			RespondValidationErrors(w, r, validationErrors)
			return
		}

		// Ensure request context is still valid before entering implementation logic.
		if err := CheckContext(r.Context()); err != nil {
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}

		// Call implementation
		result, err := impl.CreateTenant(r.Context(), &input)
		if err != nil {
			// Get logger from context and use proper error handling
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}
		// Validate output (if validation method exists)
		if validator, ok := interface{}(result).(interface {
			Validate() interface{ HasErrors() bool }
		}); ok {
			if validationErrors := validator.Validate(); validationErrors.HasErrors() {
				RespondError(w, r, gohttp.StatusInternalServerError, "Invalid response from implementation")
				return
			}
		}

		// Respond with result
		RespondJSONEnvelope(w, gohttp.StatusOK, result, r)
	}
}

// createTenantGetTenantHandler creates a handler for GET /api/tenants/{id}
//
// Fetch one tenant by id.
func createTenantGetTenantHandler(impl TenantImplementation) gohttp.HandlerFunc {
	return func(w gohttp.ResponseWriter, r *gohttp.Request) {
		// Extract path parameters
		IdStr := chi.URLParam(r, "id")
		if IdStr == "" {
			RespondError(w, r, gohttp.StatusBadRequest, "id is required")
			return
		}
		Id, err := types.ParseIdentityUUID(IdStr)
		if err != nil {
			RespondError(w, r, gohttp.StatusBadRequest, "id must be a valid UUID")
			return
		}
		// Extract query parameters
		IncludeArchived := parseBoolQueryParam(r, "includeArchived", false)

		// Ensure request context is still valid before entering implementation logic.
		if err := CheckContext(r.Context()); err != nil {
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}

		// Call implementation
		result, err := impl.GetTenant(r.Context(), Id, IncludeArchived)
		if err != nil {
			// Get logger from context and use proper error handling
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}
		// Validate output (if validation method exists)
		if validator, ok := interface{}(result).(interface {
			Validate() interface{ HasErrors() bool }
		}); ok {
			if validationErrors := validator.Validate(); validationErrors.HasErrors() {
				RespondError(w, r, gohttp.StatusInternalServerError, "Invalid response from implementation")
				return
			}
		}

		// Respond with result
		RespondJSONEnvelope(w, gohttp.StatusOK, result, r)
	}
}

// createTenantUpdateSecretHandler creates a handler for PATCH /api/tenants/{id}
func createTenantUpdateSecretHandler(impl TenantImplementation) gohttp.HandlerFunc {
	return func(w gohttp.ResponseWriter, r *gohttp.Request) {
		// Extract path parameters
		IdStr := chi.URLParam(r, "id")
		if IdStr == "" {
			RespondError(w, r, gohttp.StatusBadRequest, "id is required")
			return
		}
		Id, err := types.ParseIdentityUUID(IdStr)
		if err != nil {
			RespondError(w, r, gohttp.StatusBadRequest, "id must be a valid UUID")
			return
		}
		// Parse scalar arguments from request body (non-GET endpoints)
		var requestBody struct {
			Secret string `json:"secret"`
		}
		if err := json.NewDecoder(r.Body).Decode(&requestBody); err != nil {
			RespondError(w, r, gohttp.StatusBadRequest, "Invalid request body")
			return
		}

		// Validate secret
		if validator, ok := interface{}(requestBody.Secret).(interface {
			ValidateRequired() (bool, []types.ValidationError)
		}); ok {
			if valid, fieldErrs := validator.ValidateRequired(); !valid {
				validationErrors := types.NewValidationErrors()
				validationErrors.SetFieldErrors("secret", fieldErrs)
				RespondValidationErrors(w, r, validationErrors)
				return
			}
		}

		// Ensure request context is still valid before entering implementation logic.
		if err := CheckContext(r.Context()); err != nil {
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}

		// Call implementation
		result, err := impl.UpdateSecret(r.Context(), Id, requestBody.Secret)
		if err != nil {
			// Get logger from context and use proper error handling
			logger := LoggerFromContext(r.Context())
			RespondAppError(w, logger, err)
			return
		}
		// Validate output (if validation method exists)
		if validator, ok := interface{}(result).(interface {
			Validate() interface{ HasErrors() bool }
		}); ok {
			if validationErrors := validator.Validate(); validationErrors.HasErrors() {
				RespondError(w, r, gohttp.StatusInternalServerError, "Invalid response from implementation")
				return
			}
		}

		// Respond with result
		RespondJSONEnvelope(w, gohttp.StatusOK, result, r)
	}
}

// =============================================================================
// Query parameter parsing helpers
// =============================================================================
// These functions parse query parameters from the URL and convert them to Go types.
// They handle missing values gracefully by returning defaults or nil for pointer types.

// parseIntQueryParam parses an integer query parameter with a default value.
// Returns the default if the parameter is missing or cannot be parsed.
func parseIntQueryParam(r *gohttp.Request, name string, defaultVal int64) int64 {
	val := r.URL.Query().Get(name)
	if val == "" {
		return defaultVal
	}
	parsed, err := strconv.ParseInt(val, 10, 64)
	if err != nil {
		return defaultVal
	}
	return parsed
}

// parseIntQueryParamPtr parses an optional integer query parameter.
// Returns nil if the parameter is missing or cannot be parsed.
func parseIntQueryParamPtr(r *gohttp.Request, name string) *int64 {
	val := r.URL.Query().Get(name)
	if val == "" {
		return nil
	}
	parsed, err := strconv.ParseInt(val, 10, 64)
	if err != nil {
		return nil
	}
	return &parsed
}

// parseStringQueryParam parses a string query parameter with a default value.
// Returns the default if the parameter is missing.
func parseStringQueryParam(r *gohttp.Request, name string, defaultVal string) string {
	val := r.URL.Query().Get(name)
	if val == "" {
		return defaultVal
	}
	return val
}

// parseStringQueryParamPtr parses an optional string query parameter.
// Returns nil if the parameter is missing.
func parseStringQueryParamPtr(r *gohttp.Request, name string) *string {
	val := r.URL.Query().Get(name)
	if val == "" {
		return nil
	}
	return &val
}

// parseBoolQueryParam parses a boolean query parameter with a default value.
// Accepts: true, false, 1, 0 (case-insensitive).
// Returns the default if the parameter is missing or cannot be parsed.
func parseBoolQueryParam(r *gohttp.Request, name string, defaultVal bool) bool {
	val := r.URL.Query().Get(name)
	if val == "" {
		return defaultVal
	}
	parsed, err := strconv.ParseBool(val)
	if err != nil {
		return defaultVal
	}
	return parsed
}

// parseBoolQueryParamPtr parses an optional boolean query parameter.
// Returns nil if the parameter is missing or cannot be parsed.
func parseBoolQueryParamPtr(r *gohttp.Request, name string) *bool {
	val := r.URL.Query().Get(name)
	if val == "" {
		return nil
	}
	parsed, err := strconv.ParseBool(val)
	if err != nil {
		return nil
	}
	return &parsed
}

// parseFloat64QueryParam parses a float64 query parameter with a default value.
// Returns the default if the parameter is missing or cannot be parsed.
func parseFloat64QueryParam(r *gohttp.Request, name string, defaultVal float64) float64 {
	val := r.URL.Query().Get(name)
	if val == "" {
		return defaultVal
	}
	parsed, err := strconv.ParseFloat(val, 64)
	if err != nil {
		return defaultVal
	}
	return parsed
}

// parseFloat64QueryParamPtr parses an optional float64 query parameter.
// Returns nil if the parameter is missing or cannot be parsed.
func parseFloat64QueryParamPtr(r *gohttp.Request, name string) *float64 {
	val := r.URL.Query().Get(name)
	if val == "" {
		return nil
	}
	parsed, err := strconv.ParseFloat(val, 64)
	if err != nil {
		return nil
	}
	return &parsed
}

// =============================================================================
// Utility handler factories
// =============================================================================

func createIndexHandler() gohttp.HandlerFunc {
	return func(w gohttp.ResponseWriter, r *gohttp.Request) {
		w.Header().Set("Content-Type", "text/html; charset=utf-8")
		w.Write([]byte(indexHTML))
	}
}

func createOpenAPIHandler(apiVersion, baseURL string) gohttp.HandlerFunc {
	spec, err := injectOpenAPIRuntimeValues(openAPISpec, apiVersion, baseURL)
	if err != nil {
		spec = openAPISpec
	}

	return func(w gohttp.ResponseWriter, r *gohttp.Request) {
		w.Header().Set("Content-Type", "application/json")
		w.Write([]byte(spec))
	}
}

func injectOpenAPIRuntimeValues(specJSON, apiVersion, baseURL string) (string, error) {
	var spec map[string]interface{}
	if err := json.Unmarshal([]byte(specJSON), &spec); err != nil {
		return "", err
	}

	info, ok := spec["info"].(map[string]interface{})
	if !ok {
		info = make(map[string]interface{})
		spec["info"] = info
	}
	info["version"] = apiVersion

	spec["servers"] = []map[string]interface{}{
		{
			"url":         baseURL,
			"description": "Runtime server",
		},
	}

	updated, err := json.MarshalIndent(spec, "", "  ")
	if err != nil {
		return "", err
	}

	return string(updated), nil
}

func createRapiDocHandler() gohttp.HandlerFunc {
	return func(w gohttp.ResponseWriter, r *gohttp.Request) {
		w.Header().Set("Content-Type", "text/html; charset=utf-8")
		w.Write([]byte(rapidocHTML))
	}
}